fixed bug with the login and added validations

+++ src/main/java/com/udea/bancoudea/service/CustomerService.java
@@ -38,6 +38,14 @@ public class CustomerService {
             throw new IllegalArgumentException("Balance cannot be null");
         }
 
+        if (postCustomerDTO.getBalance() < 0) {
+            throw new IllegalArgumentException("Balance cannot be negative");
+        }
+
+        if (postCustomerDTO.getFirstName() == null || postCustomerDTO.getLastName() == null) {
+            throw new IllegalArgumentException("First name and last name cannot be null");
+        }
+
         CustomerDTO customerDTO = new CustomerDTO();
         customerDTO.setId(postCustomerDTO.getId());
         customerDTO.setBalance(postCustomerDTO.getBalance());
@@ -68,11 +76,17 @@ public class CustomerService {
     }
 
     public CustomerDTO updateCustomer(CustomerDTO customerDTO){
+        if (!customerRepository.existsById(customerDTO.getId())) {
+            throw new RuntimeException("Customer with ID " + customerDTO.getId() + " not found for update");
+        }
         Customer customer = customerMapper.toEntity(customerDTO);
         return customerMapper.toDTO(customerRepository.save(customer));
     }
 
     public void deleteCustomer(Long id){
+        if (!customerRepository.existsById(id)) {
+            throw new RuntimeException("Customer with ID " + id + " not found for delete");
+        }
         customerRepository.deleteById(id);
     }
